tests: improve CLI target tests
